nl_query/requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9
python-dotenv>=1.0
rich>=13.0
tabulate>=0.9
